Refactor: Update SVG icons and components to use new edit icon (#20650)
