+++ openmetadata-ui/src/main/resources/ui/src/components/ActivityFeed/ActivityFeedCard/PopoverContent.tsx
@@ -15,8 +15,8 @@ import Icon from '@ant-design/icons';
 import { Popover, Space } from 'antd';
 import { isNil, isUndefined, uniqueId } from 'lodash';
 import React, { FC, useMemo, useState } from 'react';
+import { ReactComponent as IconEdit } from '../../../assets/svg/edit-new.svg';
 import { ReactComponent as IconFeedDelete } from '../../../assets/svg/ic-delete.svg';
-import { ReactComponent as IconEdit } from '../../../assets/svg/ic-edit.svg';
 import { ReactComponent as IconReaction } from '../../../assets/svg/ic-reaction.svg';
 import { ReactComponent as IconReplyFeed } from '../../../assets/svg/ic-reply.svg';
 import { REACTION_LIST } from '../../../constants/reactions.constant';
