+++ openmetadata-ui/src/main/resources/ui/src/components/Entity/Task/TaskTab/task-tab-new.less
@@ -68,18 +68,11 @@
       &:hover {
         background: #f5faff;
         color: #1570ef !important;
-        .anticon svg path {
-          stroke: #1570ef;
-        }
       }
 
       &.ant-dropdown-menu-item-selected {
         background: #f5faff;
         color: #1570ef !important;
-        .anticon svg path {
-          stroke: #1570ef !important;
-          // fill:'#1570ef';
-        }
       }
 
       &:last-child {
