+++ openmetadata-ui/src/main/resources/ui/src/components/common/OwnerLabel/OwnerLabel.component.tsx
@@ -114,6 +114,7 @@ export const OwnerLabel = ({
       <div
         className={classNames({
           'owner-label-container': !isCompactView,
+          'd-flex owner-label-heading gap-2 items-center': isCompactView,
         })}
         data-testid="owner-label">
         {/* Owner avatars list */}
