+++ openmetadata-ui/src/main/resources/ui/src/utils/TasksUtils.ts
@@ -16,8 +16,8 @@ import i18Next from 'i18next';
 import { isEmpty, isEqual, isUndefined } from 'lodash';
 import React from 'react';
 import { ReactComponent as CancelColored } from '../assets/svg/cancel-colored.svg';
+import { ReactComponent as EditSuggestionIcon } from '../assets/svg/edit-new.svg';
 import { ReactComponent as CloseIcon } from '../assets/svg/ic-close-circle.svg';
-import { ReactComponent as EditSuggestionIcon } from '../assets/svg/ic-edit-suggestion.svg';
 import { ReactComponent as CheckIcon } from '../assets/svg/ic-tick-circle.svg';
 import { ActivityFeedTabs } from '../components/ActivityFeed/ActivityFeedTab/ActivityFeedTab.interface';
 import { FQN_SEPARATOR_CHAR } from '../constants/char.constants';
